+++ gd-search-common/src/main/java/de/seitenbau/govdata/search/common/cache/util/SocialMediaPlatformsConsts.java
@@ -1,0 +1,22 @@
+package de.seitenbau.govdata.search.common.cache.util;
+
+/**
+ * The class contains all constants related to social media platforms
+ * 
+ * @author youalad
+ *
+ */
+public abstract class SocialMediaPlatformsConsts
+{
+  /** Twitter */
+  public static final String TWITTER = "twitter";
+
+  /** Retweeted */
+  public static final String TWITTER_RETWEETED = "retweeted";
+
+  /** Mastodon */
+  public static final String MASTODON = "mastodon";
+
+  /** Boosted */
+  public static final String MASTODON_BOOSTED = "boosted";
+}