+++ gd-search-common/src/test/java/de/seitenbau/govdata/search/common/cache/MastodonCacheTest.java
@@ -24,37 +24,34 @@ import com.google.gson.JsonPrimitive;
 import de.seitenbau.govdata.search.common.cache.util.PostContainer;
 
 @RunWith(MockitoJUnitRunner.class)
-public class TweetCacheTest extends SocialMediaCacheBase
+public class MastodonCacheTest extends SocialMediaCacheBase
 {
-  private static final String TWITTER_TYPE_POST = "post";
+  private static final String MASTODON_TYPE_POST = "post";
 
-  private static final String TWITTER_TYPE_RETWEET = "retweet";
+  private static final String MASTODON_TYPE_BOOSTED = "boosted";
 
-  private static final String BEARER_TOKEN = "12345-67890";
+  private static final String MASTODON_ENDPOINT = "http://mastodon.endpoint";
 
-  private static final String TWITTER_ENDPOINT = "http://twitter.endpoint";
-
-  private static final String TWITTER_URL = "http://twitter.url";
+  private static final String MASTODON_URL = "http://mastodon.url";
 
   @InjectMocks
-  private TweetCache sut;
+  private MastodonCache sut;
 
   @Before
   public void setup() throws Exception
   {
-    ReflectionTestUtils.setField(sut, "twitterUrl", TWITTER_URL);
-    ReflectionTestUtils.setField(sut, "tweetEndpoint", TWITTER_ENDPOINT);
-    ReflectionTestUtils.setField(sut, "twitterApiToken", BEARER_TOKEN);
+    ReflectionTestUtils.setField(sut, "mastodonUrl", MASTODON_URL);
+    ReflectionTestUtils.setField(sut, "postEndpoint", MASTODON_ENDPOINT);
     Mockito.when(httpClient.execute(Mockito.any(HttpUriRequest.class))).thenReturn(httpResponse);
   }
 
   @Test
-  public void getTweetData_entityNull() throws Exception
+  public void getPostData_entityNull() throws Exception
   {
     /* prepare */
 
     /* execute */
-    PostContainer result = sut.getTweetData();
+    PostContainer result = sut.getPostData();
 
     /* assert */
     Assertions.assertThat(result).isNull();
@@ -63,15 +60,15 @@ public class TweetCacheTest extends SocialMediaCacheBase
   }
 
   @Test
-  public void getTweetData_responseNotParsable() throws Exception
+  public void getPostData_responseNotParsable() throws Exception
   {
     /* prepare */
     Mockito.when(httpResponse.getEntity()).thenReturn(httpEntity);
-    String contentString = "invalidJson";
+    String contentString = "[invalidJson]";
     mockHttpEntityMethodsreturnValues(contentString);
 
     /* execute */
-    PostContainer result = sut.getTweetData();
+    PostContainer result = sut.getPostData();
 
     /* assert */
     Assertions.assertThat(result).isNull();
@@ -79,7 +76,7 @@ public class TweetCacheTest extends SocialMediaCacheBase
   }
 
   @Test
-  public void getTweetData_cannotReadLatestPostFromJson() throws Exception
+  public void getPostData_cannotReadLatestPostFromJson() throws Exception
   {
     /* prepare */
     Mockito.when(httpResponse.getEntity()).thenReturn(httpEntity);
@@ -89,7 +86,7 @@ public class TweetCacheTest extends SocialMediaCacheBase
     mockHttpEntityMethodsreturnValues(contentString);
 
     /* execute */
-    PostContainer result = sut.getTweetData();
+    PostContainer result = sut.getPostData();
 
     /* assert */
     Assertions.assertThat(result).isNull();
@@ -97,98 +94,94 @@ public class TweetCacheTest extends SocialMediaCacheBase
   }
 
   @Test
-  public void getTweetData() throws Exception
+  public void getPostData() throws Exception
   {
     /* prepare */
     Mockito.when(httpResponse.getEntity()).thenReturn(httpEntity);
     String createDateString = "2022-03-29T11:43:13.671Z";
     LocalDateTime createDate =
         LocalDateTime.parse(createDateString, DateTimeFormatter.ofPattern("yyyy-MM-dd'T'HH:mm:ss.SSS'Z'"));
-    JsonObject json = createDefaultJsonResponse(createDate);
-    String contentString = new Gson().toJson(json);
-    mockHttpEntityMethodsreturnValues(contentString);
-
-    /* execute */
-    PostContainer result = sut.getTweetData();
-
-    /* assert */
-    assertSocialMediaPost(result, createDate, TWITTER_TYPE_POST, TWITTER_URL);
-    Mockito.verify(httpResponse, times(2)).getEntity();
-  }
-
-  @Test
-  public void getTweetData_referenced_tweets() throws Exception
-  {
-    /* prepare */
-    Mockito.when(httpResponse.getEntity()).thenReturn(httpEntity);
-    String createDateString = "2022-03-29T11:43:13.671Z";
-    LocalDateTime createDate =
-        LocalDateTime.parse(createDateString, DateTimeFormatter.ofPattern("yyyy-MM-dd'T'HH:mm:ss.SSS'Z'"));
-    JsonObject json = createDefaultJsonResponse(createDate);
-    JsonArray dataArray = json.getAsJsonArray("data");
-    JsonObject latestPostObject = dataArray.get(0).getAsJsonObject();
-    JsonArray refTweetArray = new JsonArray();
-    JsonObject objectRefTweets = new JsonObject();
-    objectRefTweets.add("type", new JsonPrimitive(TWITTER_TYPE_RETWEET));
-    refTweetArray.add(objectRefTweets);
-    latestPostObject.add("referenced_tweets", refTweetArray);
-    dataArray.add(latestPostObject);
-    json.add("data", dataArray);
-    String contentString = new Gson().toJson(json);
-    mockHttpEntityMethodsreturnValues(contentString);
-
-    /* execute */
-    PostContainer result = sut.getTweetData();
-
-    /* assert */
-    assertSocialMediaPost(result, createDate, TWITTER_TYPE_RETWEET, TWITTER_URL);
-    Mockito.verify(httpResponse, times(2)).getEntity();
-  }
-
-  @Test
-  public void getTweetData_createDate_notParsable() throws Exception
-  {
-    /* prepare */
-    Mockito.when(httpResponse.getEntity()).thenReturn(httpEntity);
-    JsonObject json = createDefaultJsonResponse(null);
-    String contentString = new Gson().toJson(json);
-    mockHttpEntityMethodsreturnValues(contentString);
-
-    /* execute */
-    PostContainer result = sut.getTweetData();
-
-    /* assert */
-    Assertions.assertThat(result).isNotNull();
-    assertSocialMediaPost(result, null, TWITTER_TYPE_POST, TWITTER_URL);
-    Mockito.verify(httpResponse, times(2)).getEntity();
-  }
-
-  private JsonObject createDefaultJsonResponse(LocalDateTime createDate)
-  {
-    JsonObject json = new JsonObject();
     JsonArray dataArray = new JsonArray();
+    JsonObject objectData = createDefaultJsonObject(createDate);
+    objectData.add("content", new JsonPrimitive(POST_TEXT));
+    objectData.add("reblog", null);
+    dataArray.add(objectData);
+    String contentString = new Gson().toJson(dataArray);
+    mockHttpEntityMethodsreturnValues(contentString);
+
+    /* execute */
+    PostContainer result = sut.getPostData();
+
+    /* assert */
+    assertSocialMediaPost(result, createDate, MASTODON_TYPE_POST, MASTODON_URL);
+    Mockito.verify(httpResponse, times(2)).getEntity();
+  }
+
+  @Test
+  public void getPostData_referenced_post() throws Exception
+  {
+    /* prepare */
+    Mockito.when(httpResponse.getEntity()).thenReturn(httpEntity);
+    String createDateString = "2022-03-29T11:43:13.671Z";
+    LocalDateTime createDate =
+        LocalDateTime.parse(createDateString, DateTimeFormatter.ofPattern("yyyy-MM-dd'T'HH:mm:ss.SSS'Z'"));
+    JsonArray dataArray = new JsonArray();
+    JsonObject objectData = createDefaultJsonObject(createDate);
+    objectData.add("content", null);
+    JsonObject reblogData = new JsonObject();
+    reblogData.add("content", new JsonPrimitive(POST_TEXT));
+    objectData.add("reblog", reblogData);
+    dataArray.add(objectData);
+    String contentString = new Gson().toJson(dataArray);
+    mockHttpEntityMethodsreturnValues(contentString);
+
+    /* execute */
+    PostContainer result = sut.getPostData();
+
+    /* assert */
+    assertSocialMediaPost(result, createDate, MASTODON_TYPE_BOOSTED, MASTODON_URL);
+    Mockito.verify(httpResponse, times(2)).getEntity();
+  }
+
+  @Test
+  public void getPostData_createDate_notParsable() throws Exception
+  {
+    /* prepare */
+    Mockito.when(httpResponse.getEntity()).thenReturn(httpEntity);
+    JsonArray dataArray = new JsonArray();
+    JsonObject objectData = createDefaultJsonObject(null);
+    objectData.add("content", new JsonPrimitive(POST_TEXT));
+    objectData.add("reblog", null);
+    dataArray.add(objectData);
+    String contentString = new Gson().toJson(dataArray);
+    mockHttpEntityMethodsreturnValues(contentString);
+
+    /* execute */
+    PostContainer result = sut.getPostData();
+
+    /* assert */
+    assertSocialMediaPost(result, null, MASTODON_TYPE_POST, MASTODON_URL);
+    Mockito.verify(httpResponse, times(2)).getEntity();
+  }
+
+  private JsonObject createDefaultJsonObject(LocalDateTime createDate)
+  {
     JsonObject objectData = new JsonObject();
     objectData.add("id", new JsonPrimitive(POST_ID));
-    objectData.add("text", new JsonPrimitive(POST_TEXT));
     String createDateValue = "";
     if (createDate != null)
     {
       // convert to UTC
       createDateValue =
           DateTimeFormatter.ofPattern("yyyy-MM-dd'T'HH:mm:ss.SSS'Z'").withZone(ZoneId.of("UTC"))
-          .format(createDate.atZone(ZONE_ID_BERLIN).toInstant());
+              .format(createDate.atZone(ZONE_ID_BERLIN).toInstant());
     }
     objectData.add("created_at", new JsonPrimitive(createDateValue));
-    dataArray.add(objectData);
-    json.add("data", dataArray);
-    JsonObject includesObject = new JsonObject();
-    JsonArray usersArray = new JsonArray();
-    JsonObject userObject = new JsonObject();
-    userObject.add("username", new JsonPrimitive(USER_USERNAME));
-    userObject.add("name", new JsonPrimitive(USER_NAME));
-    usersArray.add(userObject);
-    includesObject.add("users", usersArray);
-    json.add("includes", includesObject);
-    return json;
+    JsonObject accountObject = new JsonObject();
+    accountObject.add("username", new JsonPrimitive(USER_USERNAME));
+    accountObject.add("display_name", new JsonPrimitive(USER_NAME));
+    objectData.add("account", accountObject);
+    return objectData;
   }
+
 }